viikko4/TddOstoskori/src/test/java/ohtu/OstoskoriTest.java
package ohtu;

import java.util.Arrays;
import org.junit.Before;
import org.junit.Test;
import static org.junit.Assert.*;

public class OstoskoriTest {

    Ostoskori kori;
    Tuote maito;


    @Before
    public void setUp() {
        kori = new Ostoskori();
        maito = new Tuote("maito", 3);
    }

    // step 1
    @Test
    public void ostoskorinHintaJaTavaroidenMaaraAlussa() { 
        assertEquals(0, kori.hinta());
 
        assertEquals(0, kori.tavaroitaKorissa());
    }

    // step 2
    @Test
    public void yhdenTuotteenLisaamisenJalkeenKorissaYksiTuote() { 
        kori.lisaaTuote(maito);

        assertEquals(1, kori.tavaroitaKorissa());
    }

    @Test
    public void yhdenTuotteenLisaaminenKasvattaaOstoskorinHintaa() {
        kori.lisaaTuote(maito);

        assertEquals(3, kori.hinta());
    }

}
